requirements.txt
pyTelegramBotAPI==4.4.0
python-dotenv==0.19.2
peewee~=3.15.3
requests~=2.27.1
loguru~=0.6.0